% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{close_duckdb_connection}
\alias{close_duckdb_connection}
\title{Close a DuckDB connection}
\usage{
close_duckdb_connection(con)
}
\arguments{
\item{con}{The DuckDB connection to close}
}
\description{
Close a DuckDB connection
}
