% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{collect_processors}
\alias{collect_processors}
\title{Collect all processors (validators, summarizers, visualizers) for a definition}
\usage{
collect_processors(def)
}
\arguments{
\item{def}{The definition list (as returned by read_definition)}
}
\value{
A list where each element is a variable with its processors
}
\description{
Collect all processors (validators, summarizers, visualizers) for a definition
}
