% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{column_exists}
\alias{column_exists}
\title{Check if a column exists in a DuckDB table}
\usage{
column_exists(duckdb_conn, table_name, column_name)
}
\arguments{
\item{duckdb_conn}{DuckDB connection object}

\item{table_name}{Name of the table to check}

\item{column_name}{Name of the column to check}
}
\value{
Boolean indicating if the column exists
}
\description{
Check if a column exists in a DuckDB table
}
