% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_table.R
\name{create_lookup_db}
\alias{create_lookup_db}
\title{Create canonical value lookup DuckDB database}
\usage{
create_lookup_db(tables, save_location = NULL)
}
\arguments{
\item{tables}{Character vector of table names to scaffold.}

\item{save_location}{Optional file path for the DuckDB database. If NULL, creates an in-memory DB.}
}
\value{
A DBI connection to the DuckDB database.
}
\description{
Create canonical value lookup DuckDB database
}
