% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookupTable.R
\name{createLookupDb}
\alias{createLookupDb}
\title{Create canonical value lookup DuckDB database}
\usage{
createLookupDb(tables, save_location = NULL)
}
\arguments{
\item{tables}{Character vector of table names to scaffold.}

\item{save_location}{Optional file path for the DuckDB database. If NULL, creates an in-memory DB.}
}
\value{
A DBI connection to the DuckDB database.
}
\description{
Create canonical value lookup DuckDB database
}
