% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{duckdb_validator_wrapper}
\alias{duckdb_validator_wrapper}
\title{Decorator for DuckDB validators}
\usage{
duckdb_validator_wrapper(validator_fn)
}
\arguments{
\item{validator_fn}{The core validator function (conn, table_name, column_name, ...)}
}
\value{
A function that can take either a connection or a path
}
\description{
Decorator for DuckDB validators
}
