% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connection.R
\name{get_depot_connection}
\alias{get_depot_connection}
\title{Get a Database Connection}
\usage{
get_depot_connection(connection_details)
}
\arguments{
\item{connection_details}{A list of database connection details, as returned by `get_depot_connection_details()`.}
}
\value{
A `DBIConnection` object that represents the database connection.
}
\description{
This function establishes a database connection using the details provided.
}
\examples{
\dontrun{
connection_details <- get_depot_connection_details(
  host = "localhost",
  user = "myuser",
  password = "mypassword",
  dbname = "mydatabase"
)
conn <- get_depot_connection(connection_details)
DBI::dbDisconnect(conn)
}
}
