% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{get_duckdb_row_count}
\alias{get_duckdb_row_count}
\title{Get total row count from a DuckDB table}
\usage{
get_duckdb_row_count(duckdb_conn, table_name)
}
\arguments{
\item{duckdb_conn}{DuckDB connection object}

\item{table_name}{Name of the table}
}
\value{
Total number of rows
}
\description{
Get total row count from a DuckDB table
}
