% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{get_duckdb_stats}
\alias{get_duckdb_stats}
\title{Get database statistics from a DuckDB file}
\usage{
get_duckdb_stats(data_file_path, table_name = "data")
}
\arguments{
\item{data_file_path}{Path to the DuckDB data file}

\item{table_name}{Name of the table to analyze (default: "data")}
}
\value{
A list containing:
\item{row_count}{Total number of rows}
\item{column_types}{Data frame with column type information}
\item{column_stats}{Data frame with column statistics}
}
\description{
Get database statistics from a DuckDB file
}
