% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{get_type_functions}
\alias{get_type_functions}
\title{Get type-specific functions for a column}
\usage{
get_type_functions(type)
}
\arguments{
\item{type}{The type of the column}
}
\value{
List of validators, summarizers, and visualizers for the type
}
\description{
Get type-specific functions for a column
}
