% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_processors.R
\name{get_type_validators}
\alias{get_type_validators}
\title{Get the list of type-specific validators, summarizers, and visualizers}
\usage{
get_type_validators()
}
\value{
A list of type-specific functions organized by type (string, date, number)
}
\description{
Get the list of type-specific validators, summarizers, and visualizers
}
