% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{get_validator_context}
\alias{get_validator_context}
\title{Get context string for a validator (including message if present)}
\usage{
get_validator_context(validator)
}
\arguments{
\item{validator}{Validator object or string}
}
\value{
Formatted context string
}
\description{
Get context string for a validator (including message if present)
}
