% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition.R
\name{get_var_definition}
\alias{get_var_definition}
\title{Get a specific variable definition from a table definition}
\usage{
get_var_definition(table_definition, var_name)
}
\arguments{
\item{table_definition}{The table definition list}

\item{var_name}{The name of the variable to get}
}
\value{
The variable definition for the specified variable
}
\description{
Get a specific variable definition from a table definition
}
