% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnection.R
\name{getDepotConnection}
\alias{getDepotConnection}
\title{Get a Database Connection}
\usage{
getDepotConnection(connectionDetails)
}
\arguments{
\item{connectionDetails}{A list of database connection details, as returned by `getDepotConnectionDetails()`.}
}
\value{
A `DBIConnection` object that represents the database connection.
}
\description{
This function establishes a database connection using the details provided.
}
\examples{
\dontrun{
connectionDetails <- getDepotConnectionDetails(
  host = "localhost",
  user = "myuser",
  password = "mypassword",
  dbname = "mydatabase"
)
conn <- getDepotConnection(connectionDetails)
DBI::dbDisconnect(conn)
}
}
