% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_file_summary}
\alias{load_file_summary}
\title{Load an object from an RDS file}
\usage{
load_file_summary(file_path)
}
\arguments{
\item{file_path}{The path to the RDS file.}
}
\value{
The loaded R object.
}
\description{
Load an object from an RDS file
}
