% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cache.R
\name{loadFileSummary}
\alias{loadFileSummary}
\title{Load an object from an RDS file}
\usage{
loadFileSummary(file_path)
}
\arguments{
\item{file_path}{The path to the RDS file.}
}
\value{
The loaded R object.
}
\description{
Load an object from an RDS file
}
