% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{open_duckdb_connection}
\alias{open_duckdb_connection}
\title{Open a DuckDB connection}
\usage{
open_duckdb_connection(data_file_path, read_only = TRUE)
}
\arguments{
\item{data_file_path}{Path to the DuckDB file}

\item{read_only}{Whether to open in read-only mode}
}
\value{
A DuckDB connection
}
\description{
Open a DuckDB connection
}
