% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{run_summary}
\alias{run_summary}
\title{Helper function to run summary operations}
\usage{
run_summary(
  duckdb_conn,
  query,
  var,
  summarizer_name,
  results_fn,
  message_fn,
  status_fn,
  total_rows
)
}
\arguments{
\item{duckdb_conn}{DuckDB connection object}

\item{query}{SQL query to execute}

\item{var}{Name of the variable being summarized}

\item{summarizer_name}{Name of the summarizer function}

\item{results_fn}{Function to process the query results}

\item{message_fn}{Function to generate summary message}

\item{status_fn}{Function to determine summary status}

\item{total_rows}{Total number of rows in the table}
}
\value{
A list containing the summary results
}
\description{
Helper function to run summary operations
}
