% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_file_summary}
\alias{save_file_summary}
\title{Save an object to an RDS file}
\usage{
save_file_summary(object, file_path)
}
\arguments{
\item{object}{The R object to save.}

\item{file_path}{The path where the RDS file should be saved.}
}
\description{
Save an object to an RDS file
}
