% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cache.R
\name{saveFileSummary}
\alias{saveFileSummary}
\title{Save an object to an RDS file}
\usage{
saveFileSummary(object, file_path)
}
\arguments{
\item{object}{The R object to save.}

\item{file_path}{The path where the RDS file should be saved.}
}
\description{
Save an object to an RDS file
}
