% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_barchart.R
\name{summarize_barchart}
\alias{summarize_barchart}
\title{Summarize variable values as a bar chart}
\usage{
summarize_barchart(duckdb_conn, table_name = "data", var, params = NULL)
}
\arguments{
\item{duckdb_conn}{DuckDB connection}

\item{table_name}{Name of the table}

\item{var}{Name of the variable to summarize}

\item{params}{Additional parameters (not used)}
}
\value{
List containing status, message, and results with ggplot object
}
\description{
Summarize variable values as a bar chart
}
