% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition.R
\name{summarize_definition}
\alias{summarize_definition}
\title{Summarize a definition (return data frame of variable names and types)}
\usage{
summarize_definition(def)
}
\arguments{
\item{def}{The definition list (as returned by read_definition)}
}
\value{
A data.frame with columns: name, type, description
}
\description{
Summarize a definition (return data frame of variable names and types)
}
