% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric_coercion.R
\name{summarize_numeric_coercion}
\alias{summarize_numeric_coercion}
\title{Analyze what percentage of string values can be coerced to numbers}
\usage{
summarize_numeric_coercion(duckdb_conn, table_name, var, params = NULL)
}
\arguments{
\item{duckdb_conn}{DuckDB connection object}

\item{table_name}{Name of the table containing the variable}

\item{var}{Name of the column to analyze}

\item{params}{Optional parameters (not used for numeric coercion analysis)}
}
\value{
A list containing:
\item{status}{"success" or "error"}
\item{message}{Description of any error}
\item{results}{A list containing:
\item{coercion_stats}{Statistics about numeric coercion:
\item{can_coerce}{Percentage of values that can be coerced to numbers}
\item{cannot_coerce}{Percentage of values that cannot be coerced to numbers}
\item{description}{Explanation of the coercion statistics}
}
}
}
\description{
Analyze what percentage of string values can be coerced to numbers
}
