% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{summarize_validators}
\alias{summarize_validators}
\title{Summarize all validators in a definition, including parameters}
\usage{
summarize_validators(def)
}
\arguments{
\item{def}{The definition list (as returned by read_definition)}
}
\value{
A data.frame with columns: name, type, validator, context
}
\description{
Summarize all validators in a definition, including parameters
}
